build: migrate bundling from rollup to rolldown
Rolldown replaces rollup plus six of its plugins with built-ins: TypeScript
transforms, node resolution, CommonJS, JSON, aliasing, and replacement all
come from the bundler now. Only rollup-plugin-copy stays.

Full bundle step drops 3.1s -> 0.85s, and watch mode rebuilds the V8 adapter
in ~40ms instead of a full pass.

One behavior change fell out of the resolver. `pkce-challenge` exposes no bare
"default" export condition, which rollup's resolver could not handle, so the
package was left external and a renderChunk string-replace deleted the leftover
import. Rolldown resolves it and bundles it instead. Since the portal only
talks to a local HTTP server, the MCP SDK's OAuth path is dead code, so this
aliases in a stub the same way the code-exec flags do. That keeps the
dependency out of the bundle deliberately rather than as a side effect of a
resolver failure.

terser was doing no minification (compress and mangle both off) -- only comment
stripping and pretty-printing. `output.comments: false` covers it.

Verified: the built portal's MCP initialize and tools/list responses are
byte-identical to the rollup build, and the release/debug feature-flag bundles
still differ in the same places.
