+++ package.json
@@ -31,10 +31,10 @@
     "ai-music-production"
   ],
   "scripts": {
-    "build": "npm run parser:build && npm run ui:build && rollup -c config/rollup.config.mjs && npm run dxt:build",
+    "build": "npm run parser:build && npm run ui:build && rolldown -c config/rolldown.config.mjs && npm run dxt:build",
     "build:debug": "ENABLE_LIVE_API=true ENABLE_CODE_EXEC=true ENABLE_WARP_MARKERS=true npm run build",
-    "build:watch": "npm run build && rollup -c config/rollup.config.mjs -w",
-    "dev": "npm run parser:build && rollup -c config/rollup.config.mjs -w",
+    "build:watch": "npm run build && rolldown -c config/rolldown.config.mjs -w",
+    "dev": "npm run parser:build && rolldown -c config/rolldown.config.mjs -w",
     "dev:debug": "ENABLE_LIVE_API=true ENABLE_CODE_EXEC=true ENABLE_WARP_MARKERS=true npm run dev",
     "test": "npm run test:setup && vitest run --reporter=./config/vitest-minimal-reporter.mjs --silent=true",
     "test:watch": "npm run test:setup && vitest --reporter=dot --silent=true",
@@ -139,12 +139,6 @@
     "@eslint-community/eslint-plugin-eslint-comments": "4.7.2",
     "@playwright/test": "1.62.1",
     "@preact/preset-vite": "2.10.6",
-    "@rollup/plugin-alias": "6.0.0",
-    "@rollup/plugin-commonjs": "29.0.3",
-    "@rollup/plugin-json": "6.1.0",
-    "@rollup/plugin-node-resolve": "16.0.3",
-    "@rollup/plugin-replace": "6.0.3",
-    "@rollup/plugin-terser": "1.0.0",
     "@stryker-mutator/core": "9.6.1",
     "@stryker-mutator/vitest-runner": "9.6.1",
     "@stylistic/eslint-plugin": "5.10.0",
@@ -165,9 +159,8 @@
     "oxlint": "1.77.0",
     "oxlint-tsgolint": "7.0.2001",
     "peggy": "5.1.0",
-    "rollup": "4.62.4",
+    "rolldown": "1.2.1",
     "rollup-plugin-copy": "3.5.0",
-    "rollup-plugin-esbuild": "6.2.1",
     "typescript": "npm:@typescript/typescript6@6.0.2",
     "vite": "8.2.1",
     "vite-plugin-singlefile": "2.3.3",
