+++ src/test/mocks/mock-chat-ui-html.ts
@@ -4,6 +4,6 @@
 
 /**
  * Mock chat UI HTML for testing.
- * The actual HTML is inlined by the Rollup plugin during the build process.
+ * The actual HTML is inlined by the Rolldown plugin during the build process.
  */
 export default "<html><body>Mock Chat UI</body></html>";
