+++ src/tools/clip/code-exec/code-executor-disabled.ts
@@ -5,11 +5,11 @@
 
 /**
  * Disabled stub for code-executor.ts.
- * Substituted by rollup when ENABLE_CODE_EXEC is not set.
+ * Substituted by rolldown when ENABLE_CODE_EXEC is not set.
  * Exports the same interface but always returns an error.
  *
  * IMPORTANT: If this file is renamed or moved, update the alias entry in
- * config/rollup.config.mjs (codeExecAliases) to match.
+ * config/rolldown.config.mjs (codeExecAliases) to match.
  */
 
 import { type SandboxResult } from "#src/tools/clip/code-exec/code-exec-types.ts";
